requirements.txt
Flask==3.1.2
GitPython==3.1.40
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
click==8.3.0
blinker==1.9.0
